% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_calc.R
\name{correct_ph_temp}
\alias{correct_ph_temp}
\title{Correct measured pH for temperature}
\usage{
correct_ph_temp(pH, temp1, temp2, sal = 35, alk = 2400, n_max = 100,
  marg = 0.01)
}
\arguments{
\item{pH}{measured pH}

\item{temp1}{temperature at which pH was measured}

\item{temp2}{in-situ temperature from samples origin}

\item{sal}{salinity of sample}

\item{alk}{measured alkalinity}

\item{n_max}{maximum number of iterations}

\item{marg}{fineness of result}
}
\description{
Correct measured pH for temperature
}
