% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_constants.R
\name{make_carb_rate}
\alias{make_carb_rate}
\title{Compute all the carbon chem rate constants}
\usage{
make_carb_rate(temp, sal)
}
\arguments{
\item{temp}{temperature of sample}

\item{sal}{}
}
\description{
Compute all the carbon chem rate constants
}
