% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_constants.R
\name{make_k0}
\alias{make_k0}
\title{Compute k0}
\usage{
make_k0(temp, sal, degc = TRUE)
}
\arguments{
\item{degc}{}
}
\description{
Compute the equalibrium constant for carbond dioxide to carbonic acid.
}
