% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_constants.R
\name{make_k1}
\alias{make_k1}
\title{Compute k1}
\usage{
make_k1(temp, sal, degc = TRUE)
}
\arguments{
\item{degc}{}
}
\description{
Compute the equalibrium constant k1 for the conversion of carbonic acid to bicarbonate
}
