% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_constants.R
\name{make_k2}
\alias{make_k2}
\title{Compute k2}
\usage{
make_k2(temp, sal, degc = TRUE)
}
\arguments{
\item{degc}{}
}
\description{
Compute the equalibrium constant k1 for the conversion of bicarbonate to carbonate
}
