% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_constants.R
\name{make_kb}
\alias{make_kb}
\title{Compute Kb}
\usage{
make_kb(temp, sal, degc = TRUE)
}
\arguments{
\item{degc}{}
}
\description{
Compute the equalibrium constant for the dissociation of boric acid to borate
}
