% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_calc.R
\name{pCO2_from_pH_dic}
\alias{pCO2_from_pH_dic}
\title{Calculate pCO2 from pH and DIC}
\usage{
pCO2_from_pH_dic(pH, dic, temp, sal)
}
\arguments{
\item{sal}{}
}
\description{
Calculate pCO2 from pH and DIC
}
