% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_calc.R
\name{pH_from_alk_dic}
\alias{pH_from_alk_dic}
\title{Calculate pH from Alkalinity and DIC}
\usage{
pH_from_alk_dic(alk, dic, temp, sal, pH_guess = 8, n_max = 100,
  marg = 0.01)
}
\arguments{
\item{pH_guess}{}
}
\description{
Calculate pH from Alkalinity and DIC
}
